Plot the top losses for a particular class
Frequently, I find useful to see the top losses (and all their visualizations) for a particular class. So, I have added a new input parameter `pclass` and a little of code to filter the indexes of the images. 

I have also added a few lines to not show the remaining axes when the square grid does not correspond to the number of images.

I have not modified the alpha value, but I consider an alpha=0.4 better to visualize the heatmap and not distort the original image .

PD. It is my first PR :)